Add nmap package
This adds nmap 7.31 as a package. I have tested the package on a Samsung Series 3 (ARM) machine and everything compiles successfully. The resulting binary works without issues. I did not add "openssl" as a dependency - this is on purpose. For a few corner cases, nmap can use some SSL functionality, but it works perfectly without. However, if the user has "openssl" installed at the time this package is being installed, then nmap will use this.