Add NcFTP (client) package
This adds NcFTP client 3.2.6 as a package. This is useful, since we don't have any standard FTP console client on ChromeOS. If the user has ncurses or openssl libraries installed, then the client can be run via an ncurses or with encryption respectively, but this is optional and therefore not defined in the dependencies. Tested the package successfully on a Samsung Series 3 (ARM) Chromebook.