Add package for bacon (BASIC-to-C converter) tool
This adds the BaCon BASIC-to-C converter (3.5). BaCon allows one to write programs using a modern BASIC dialect and generate C code and and binaries as output. The final step of the installation script fails to create a directory containing a sample source file but the tool itself works properly. Will address the example creation step in the future as time permits.

Tested sucessfully on Samsung Chromebook 3 (XE500C13-K01US).