Add package for libffi
This adds libffi (3.0.13) which provides a portable foreign-function interface library. I'm adding this in preparation for an update to the newLISP package adding ffi and proper readline support which will enhance usability of the REPL and allow use of the GUI library if you install Oracle's JRE.

Tested sucessfully on Samsung Chromebook 3 (XE500C13-K01US).