Add di (disk information) package
This adds the disk information utility as a package. 'di' is a disk information utility, displaying everything (and more) that your 'df' command does. It features the ability to display your disk usage in whatever format you prefer. It also checks the user and group quotas, so that the user sees the space available for their use, not the system wide disk space.

Tested as working properly on Samsung XE50013-K01US.