Add powerstat (0.02.11) package
Powerstat measures the power consumption of a laptop using the ACPI battery information. The output is like vmstat but also shows power consumption statistics. At the end of a run, powerstat will calculate the average, standard deviation and min/max of the gathered data.

Tested as working properly on Samsung XE50013-K01US.