Refactor logic for extracting archives
Extracting files from the downloaded archive were handled differently depending on their type. Tarballs ran tar twice; once to extract the files, and once to test for the containing directory name. Tarballs that had a file and a directory at their top level were causing problems with determining the topdir for the build step. This changes the archive extraction to work the same way for both zip and tar.

This fixes #394.