Add iftop (0.17) package
iftop does for network usage what top(1) does for CPU usage. It listens
to network traffic on a named interface and displays a table of current
bandwidth usage by pairs of hosts. Handy for answering the question "why
is our ADSL link so slow?".

Tested as working on Samsung XE50013-K01US.