Add libassuan (2.4.3) package
The libassuan package contains an inter process communication library
used by some of the other GnuPG related packages. libassuan's primary
use is to allow a client to interact with a non-persistent server.
libassuan is not, however, limited to use with GnuPG servers and
clients. It was designed to be flexible enough to meet the demands of
many transaction based environments with non-persistent servers.

Tested as working on Samsung XE50013-K01US. All tests passing.
https://gist.github.com/cstrouse/c7ef8478ca5381ca141bf7f2cebfe4d8