Update sshfs from 2.8 to 2.10
This is a bugfix and general maintenance release.

The project recommends updating to the 3.x line which requires updates to fuse
as well as creating packages for the meson build system and adding ninja as a
dependency at build time. This update will be forthcoming as I have time or
someone else tackles this.

Tested as working on XE500C13-K01US.