Download single file instead of archive
We only require the shellscript named compressdoc but have been
downloading a 1.9M archive and then copying the file into our bin
directory.

This changes the behavior to download the one file that we need via
Github's raw user content service.

It also simplifies the installation to a single line using the install
command.