Disable MPI assembler modules for ARMv8
The included MPI assembler code doesn't work for the ARMv8 processor in
the Samsung Chromebook Plus (and possibly others). Until upstream adds
support for these processors they must be disabled for the package to
build properly.

In the future support will probably be added for these processors
enabling us to re-enable support for MPI ASM modules.