Clarify prereqs.  Wiki just gets you in dev mode (#2209)
When I read this the first time, I didn't see anything on the given page telling me anything about crew. Finally I realized that although the instruction was in a separate paragraph, it didn't expect me to have already gotten in to dev mode.
This way of phrasing it should help people know what they're looking for on the wiki.