add libuuid package (#2889)
Thank you for the new package @etbusch!  I think we had this package at one time in the past and removed it since it provided the same functionality as the util_linux package.  I don't quite remember if it caused a conflict or not.  We can try it again and see what happens.