Speed up dependency checking
Previously, a lot of time on install was spent in expand_dependencies
redundantly re-examining packages and checking their dependencies.
Keeping a list that tracks which packages have already been checked for
dependencies helps avoid a lot of the extra wasted effort.

The extra list checking can be made very efficient by using Ruby's Set
class, which enforces element uniqueness and uses hash table lookups
under the hood. Set is in the Ruby stdlib but it does add one additional
require to the code.