Faster dep checks (#3106)
* Check package dependencies more efficiently

Keep track of the names of dependencies that have already been checked
in the expand_dependencies/push_dependencies loop. A quick and dirty way
to break out of dependency cycles and speed up the dependency resolution
step in general.

* Simplify the logic of dependency checking in expand_dependencies

Accomplishes the same goal without needing to explicitly keep a list of
previously-checked dependencies.