Switch to upstream installer sha384 checksum (#3843)
This replaces the hardcoded sha384 checksum for the installer to use the
sha384 checksum provided by the upstream project. Their packaging
guidelines recommend this as the preferred way of validating the
integrity of the installer. This removes the need to constantly update
our package when they make changes to the installer.