Add php80 package
This adds a first pass at a PHP 8 package.

Currently building with libressl requires a patch due to libressl missing OCB
support (even though they claim to be compatible with openssl 1.1 which includes
OCB). Once a final fix is merged upstream we can remove the patch.

Tested on x86_64. Should be tested on all other architectures prior to merge.